frontend: abstracted the actual simulation code to its own component SandboxSimulation page is now only responsible for connecting to the simulation
