+++ frontend/src/pages/SandboxSimulation/SandboxSimulation.scss
@@ -3,78 +3,4 @@
   width: 100%;
   height: 100%;
   overflow: hidden;
-
-  &--body {
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-    height: 100%;
-  }
-
-  &--toolbox {
-    flex: 0 0 content;
-  }
-
-  &--board-container {
-    display: flex;
-    flex: 1 1;
-    overflow: auto;
-  }
-
-  &--board {
-    position: relative;
-    flex: 0 0 auto;
-    width: 1600px;
-    height: 800px;
-    margin: auto;
-    border: 1px #ccc solid;
-  }
-
-  &--sliding-panel {
-    $width: 200px;
-
-    position: absolute;
-    right: -$width;
-    bottom: 100px;
-    width: $width;
-    background-color: #fff;
-    transition: 0.25s;
-
-    &:hover {
-      right: 0;
-      transition: 0.25s;
-    }
-
-    &--handle {
-      $width: 50px;
-
-      position: absolute;
-      left: -$width;
-      display: flex;
-      align-content: center;
-      align-items: center;
-      justify-content: center;
-      justify-items: center;
-      width: $width;
-      height: 100%;
-      padding: 10px;
-      background-color: #fff;
-      border: 1px solid rgb(180, 180, 180);
-      border-right: 0;
-      border-radius: 5px 0 0 5px;
-    }
-
-    &--body {
-      padding: 10px;
-      border: 1px solid rgb(180, 180, 180);
-      // border-radius: 5px 0 0 5px;
-    }
-  }
-
-  &--time-scale {
-    // specificity hack: apply the class twice
-    &#{&} {
-      width: 125px;
-    }
-  }
 }
